fixed odo create allowed  to run multiple times in same dir

+++ pkg/odo/cli/component/create.go
@@ -369,6 +369,10 @@ func (co *CreateOptions) Complete(name string, cmd *cobra.Command, args []string
 		return errors.New("this directory already contains a component")
 	}
 
+	if util.CheckPathExists(EnvFilePath) && util.CheckPathExists(co.DevfilePath) {
+		return errors.New("this directory already contains a component")
+	}
+
 	if util.CheckPathExists(EnvFilePath) && !util.CheckPathExists(co.DevfilePath) {
 		log.Warningf("Found a dangling env file without a devfile, overwriting it")
 		if err := util.DeletePath(EnvFilePath); err != nil {
